extension of ektorp to support lucene index quireies.
Since cloudant geospatian was not possible to use due to we needed a
sort order we will ahve to realy on the lucene queries which are
supported.

This will allow us to match events within a recangular box. We can then
filter out those not within the circle in java. That the only solution I
can see, but its not a big problem.

I have written a few extension classes to ektorp to add support for
lucene queries.